Built artifacts and download them for faster testing
Attempt at large speedup when buildroot doesn't need full rebuild.

When pushing to a branch named "test", it will download the latest
artifact from branch name "main" and only run the test script boot
script at the end. This give a significant speed up as a full build
and test run takes 1h20m, while just running the test takes just 12m.

This is not super useful as is but should be a good basis for
something later like just rebuilding the kernel or qemu and retesting
from that.

Signed-off-by: Michael Neuling <mikey@neuling.org>