Use caching rather than artifacts
This removes the artifact upload/download mechanism and replaces it
with caching the buildroot cccahe.

With this, the build time goes from 1h27m to 32m.

The artifact mechanism didn't work as just grabbing the output
directory causes issues with the buildroot dependency system. For
example, if qemu changes, it doesn't actually rebuild qemu on a simple
"make".

Signed-off-by: Michael Neuling <mikey@neuling.org>